feat: Renamed classes to use specific nouns about their purpose.
