+++ src/Runnable.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace StellarWP\Telemetry;
+
+interface Runnable {
+	public function run(): void;
+}
