+++ tests/wpunit/RegisterSiteTest.php
@@ -2,14 +2,14 @@
 
 use Codeception\TestCase\WPTestCase;
 use lucatume\DI52\Container;
+use StellarWP\Telemetry\Runnable;
 use StellarWP\Telemetry\ActivationHook;
-use StellarWP\Telemetry\DefaultActivationHook;
-use StellarWP\Telemetry\DefaultOptinTemplate;
-use StellarWP\Telemetry\DefaultPluginStarter;
-use StellarWP\Telemetry\DefaultTelemetryProvider;
-use StellarWP\Telemetry\PluginStarter;
+use StellarWP\Telemetry\OptInTemplate;
+use StellarWP\Telemetry\ExampleStarter;
+use StellarWP\Telemetry\DebugDataProvider;
+use StellarWP\Telemetry\Starter;
 use StellarWP\Telemetry\RegisterSiteRequest;
-use StellarWP\Telemetry\TelemetryProvider;
+use StellarWP\Telemetry\DataProvider;
 
 class RegisterSiteTest extends WPTestCase {
 	/**
@@ -26,21 +26,21 @@ class RegisterSiteTest extends WPTestCase {
 
 		// Your set up methods here.
 		$this->container = new Container();
-		$this->container->bind( TelemetryProvider::class, DefaultTelemetryProvider::class );
+		$this->container->bind( DataProvider::class, DebugDataProvider::class );
 		$this->container->singleton(
-			PluginStarter::class,
+			Starter::class,
 			function () {
-				return new DefaultPluginStarter(
-					new DefaultOptinTemplate(),
-					$this->container->get( TelemetryProvider::class )
+				return new ExampleStarter(
+					new OptInTemplate(),
+					$this->container->get( DataProvider::class )
 				);
 			},
 			[ 'init' ]
 		);
-		$this->container->singleton( ActivationHook::class, DefaultActivationHook::class );
+		$this->container->singleton( Runnable::class, ActivationHook::class );
 
 		// Run the activation hook code to register our plugin option.
-		$this->container->get( ActivationHook::class )->run();
+		$this->container->get( Runnable::class )->run();
 	}
 
 	public function tearDown(): void {
@@ -56,18 +56,23 @@ class RegisterSiteTest extends WPTestCase {
 
 		$this->assertInstanceOf( WP_Post::class, $post );
 
-		$this->assertInstanceOf( DefaultPluginStarter::class, $this->container->get( PluginStarter::class ) );
+		$this->assertInstanceOf( ExampleStarter::class, $this->container->get( Starter::class ) );
 	}
 
 	public function test_it_saves_token_on_register_site_request() {
-		$starter = $this->container->get( PluginStarter::class );
-		$request = new RegisterSiteRequest( $starter, $this->container->get( TelemetryProvider::class ) );
+		$starter = $this->container->get( Starter::class );
+		$request = new RegisterSiteRequest( $starter, $this->container->get( DataProvider::class ) );
 
 		// Test we currently have no token.
 		$this->assertArrayNotHasKey( 'token', $starter->get_meta() );
 
+		// Run the request
+		$request->run();
+
+		// Test the request was successful.
+		$this->assertArrayHasKey( 'status', $request->response );
+
 		// Test we save the user token on run.
-		$this->assertTrue( $request->run() );
 		$this->assertArrayHasKey( 'token', $starter->get_meta() );
 	}
 }
